Filtering/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
Django==3.2.12
django-filter==21.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
